moved utils to a seperate folder
